+++ code/src/Plexdata.LogWriter.Stream.Tests/Facades/StreamLoggerFacadeTests.cs
@@ -28,7 +28,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Linq;
 using System.Text;
 
 namespace Plexdata.LogWriter.Stream.Tests.Facades